Add similar 100ms delay to paste event that is in drop event;
Without the delay, in some environments, it appeared that the onPaste() code was run too early, the image was not added to the DOM element, and thus it was not successfully removed if it was forbidden one (svg for example). Adding the delay fixes the problem, or at least now it works in similar way than drag drop.