Add a 10ms slowmo to the tests
This is a bit hacky, but several tests are flaky due to race conditions
and the HTML manipulation we do in the editor. Adding a 10ms slowmo does
not slow test running in a noticeable way, but is enough to mitigate the
flakiness. And it is significantly easier than going through all flaky
tests and individually coming up with solutions to their particular
flaky parts