Fix onAnswerChange not accepting empty input
Usually the textarea value is at least a <br /> tag, but in a
test case in exam engine it was sometimes just an empty string.

Since the text field should be emptyable (some browers force a line break
in a textarea minimum content) we should also accept an empty string
in onAnswerChange.