Fix uncaught errors in Chrome onBlur
When pressing enter in math editor field, onBlur is called twice: first
in mathQuill onEnter, then the element onBlur. This didn't happen in Firefox.

The side effect was, that an error was thrown in state onblur function when
it tried to remove the stub element. Remove elements in onBlur and ignore
any raised errors.

This didn't happen when blurring in other ways.